+++ src/renderer/src/config/ocr.ts
@@ -1,4 +1,4 @@
-import { BuiltinOcrProvider, ImageOcrProvider, OcrProviderCapability } from '@renderer/types/ocr'
+import { BuiltinOcrProvider, ImageOcrProvider, OcrProviderCapability } from '@renderer/types'
 
 const tesseract: BuiltinOcrProvider & ImageOcrProvider = {
   id: 'tesseract',
@@ -10,6 +10,6 @@ const tesseract: BuiltinOcrProvider & ImageOcrProvider = {
 
 export const BUILTIN_OCR_PROVIDERS: BuiltinOcrProvider[] = [tesseract] as const
 
-export const DEFAULT_OCR_PROVIDER: Record<OcrProviderCapability, BuiltinOcrProvider> = {
+export const DEFAULT_OCR_PROVIDER: Record<OcrProviderCapability, BuiltinOcrProvider & ImageOcrProvider> = {
   image: tesseract
 } as const
