+++ src/renderer/src/utils/ocr.ts
@@ -1,5 +1,5 @@
 import TesseractLogo from '@renderer/assets/images/providers/Tesseract.js.png'
-import { isBuiltinOcrProviderId } from '@renderer/types/ocr'
+import { isBuiltinOcrProviderId } from '@renderer/types'
 
 export function getOcrProviderLogo(providerId: string) {
   if (isBuiltinOcrProviderId(providerId)) {
